feat(ucmc-web): /gazette detail page with inline PDF viewer + layout fix
Wraps the Gazette list + detail under a shared layout so the typed
Link from IssueCard's "View" button actually swaps the rendered
content. Previously gazette.tsx was the list page itself; with
gazette.$publicId.tsx as a sibling, TanStack's file router treats
the detail as a CHILD of the list and the parent has no Outlet —
clicking View updated the URL but left the list visible.

Restructuring matches the existing gear.* pattern:
  - gazette.tsx (new): minimal layout route, just <Outlet />.
  - gazette.index.tsx (renamed from gazette.tsx): the list page,
    path "/gazette/".
  - gazette.$publicId.tsx: the detail page, mounts into the parent's
    Outlet.

Detail page UI:
  - Inline <iframe> embedding the PDF from the R2 custom domain.
    Falls back to a "Download the PDF" link inside the iframe for
    browsers that can't render PDFs inline (iOS Safari sometimes).
  - "Download PDF" primary button (<a download>) and "Open in new
    tab" secondary button next to it for users who'd rather skip the
    iframe entirely.
  - Back-to-archive Link at the top.

CSP (src/server/headers.server.ts):
  - frame-src gains both R2 custom-domain hosts:
    cdn.ucmc.spencerwill.com (prod) and
    cdn.dev.ucmc.spencerwill.com (dev). Listed unconditionally so
    the same CSP string ships to both envs; the deployment branch
    determines which one resolves. 'self' alone wouldn't cover them
    — the CDN is cross-origin from the app even though same TLD.
